requirements.txt
mistralai
flask
streamlit
python-dotenv